Modify the `environment.yml` file
Add `pip` and `pip install gym` function to the `.yml` file.
This will automatically install the `gym` module when we excute the
conda create command from a file.

Also the prefix of the conda environment is modified to be hidden to
distinct from the `envs` of the `fym' module. Associated change also
included in the `.gitignore` file.