Data `model.pth` and related plots
The constraint function was wrong as it was a sigmoid which don't have a
value zero at the constraint boundary.
It was the cause of the positve 0.3 of the value function at the
altitude near zero.

This script is run on the work station.

Also, this commit will add a plotting method.

With a `PlotVar` class, a plot can be easily drawed from simple
statemetns, such as

```
    draw_safe_value(
        name='gamma-and-psi.png', model_file='model.pth',
        pvars=(gamma, psi), fvars=(z, V), fvals=(-2.4, 8))
```