Fix the optimization problem
Summary: Increase the `eps` value.

Problem: Optimization procedures usually return the initial value which
is different from intuition.

Solved: The cause is too low `eps` values which are used in numerical
differentiation for obtaining the jacobian of the `SLSQP` method. Too
small `eps` values do not produce different outputs which leads to the
zero jacobian. This can be resolved with large enough value of `eps`
which is in this case `1e-4`.